Rename 'usfmjsToSlate.js' to 'usfmToSlate.js'
